third_party/blink/web_tests/wpt_internal/css/css-conditional/container-queries/scroll-state/at-container-scroll-direction-serialization.html

<!DOCTYPE html>
<title>CSS Container Queries: scroll-state(scroll-direction) conditionText serialization</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-queries">
<link rel="help" href="https://drafts.csswg.org/cssom/#serialize-a-css-rule">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/6400">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<style id="testSheet">
  @container scroll-state(   scroll-direction:any   ) { }
  @container scroll-STate(scroll-direction:    ) { }
  @container scroll-STate(scroll-direction) { }
  @container  scroll-state(  ( scroll-Direction: LEFT) OR ( Scroll-direction: BOTTOM  ) ) { }
  @container scroll-state (scroll-direction: right) { }
</style>
<script>
  setup(() => {
    assert_implements_scroll_state_container_queries();
    assert_equals(testSheet.sheet.cssRules.length, 5);
  });

  const tests = [
      ["scroll-state(scroll-direction: any)", "Normalize spaces"],
      ["scroll-STate(scroll-direction:    )", "No value - invalid, serializes as <general-enclosed>"],
      ["scroll-state(scroll-direction)", "Boolean context"],
      ["scroll-state((scroll-direction: left) or (scroll-direction: bottom))", "Logical with 'or'"],
      ["scroll-state (scroll-direction: right)", "Not a scroll-state function with space before '('"]
  ].map((e, i) => [testSheet.sheet.cssRules[i], ...e]);

  tests.forEach((t) => {
    test(() => assert_equals(t[0].conditionText, t[1]), t[2]);
  });
</script>
